Add files via upload
About this Dataset: Netflix is one of the most popular media and video streaming platforms. They have over 8000 movies or tv shows available on their platform, as of mid-2021, they have over 200M Subscribers globally. This tabular dataset consists of listings of all the movies and tv shows available on Netflix, along with details such as - cast, directors, ratings, release year, duration, etc.

Description of each column in the dataset:

show_id: A unique identifier for each show or movie.

type: The type of content, either "Movie" or "TV Show."

title: The title of the movie or TV show.

director: The director of the movie or TV show. In the first and third entries, this information is not available (NaN).

cast: The cast or actors in the movie or TV show. In the first entry, this information is not available (NaN). In the second entry, there is a list of actors from the TV show "Blood & Water."

country: The country where the movie or TV show was produced or is associated with.

date_added: The date when the content was added to the streaming platform, in the format "Month Day, Year."

release_year: The year the movie or TV show was originally released.

rating: The content's rating, which indicates the recommended audience age or maturity level (e.g., "PG-13" or "TV-MA").

duration: The duration of the movie or TV show. In the first entry, the duration is given in minutes ("90 min"). In the second and third entries, it's indicated in the number of seasons ("2 Seasons" and "1 Season").

listed_in: The genre or category of the content, which can help classify it (e.g., "Documentaries," "International TV Shows," "Crime TV Shows").

description: A brief description or synopsis of the movie or TV show, providing an overview of the plot or subject matter.